improve decomposition of exp and sprod (#8133)
**Context:**

We've been seeing some ecosystem failures due to improperly handling the
decomposition of exponentials and scalar products.

Blocks lightning PR
https://github.com/PennyLaneAI/pennylane-lightning/pull/1246

**Description of the Change:**

Adds a `SProd` decomposition when the coefficient if of norm 1.

Simplifies the target op of an `Exp` operator prior to decomposition.

**Benefits:**

Can handle more complicated target ops in `Select` and `PrepSelPrep`
without error.

**Possible Drawbacks:**

**Related GitHub Issues:**