Fall back to op.decompose if op is unsolved in decomposition graph (#8156)
**Context:**
The DecompositionGraph currently errors out completely if any of the
operators is unsolved for, and as a result, with graph enabled, if a
circuit contains a single operator/template that is not yet integrated
with the new decomposition system, the entire thing fails, and the user
is forced to turn off graph mode and use the old decomposition system.
We want to change this behaviour, make it so that the DecompositionGraph
raise a warning instead of an error if certain operators are unsolved
for, and fall back to using op.decomposition for those operators.

**Description of the Change:**
- Downgrade error to a warning when certain operators are unsolved in
the graph.
- Add sensible warning for when the graph failed to find a solution due
to GlobalPhase.
- When the new system is enabled, operators without a decomposition will
be left undecomposed with a warning.

**Benefits:**
- The user can use the new system for the rest of the circuit even if it
contains templates that are not yet integrated with the new graph-based
system.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-98390]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>