Stepwise tape execution for MBQC (#7997)
**Context:**

We need the backend to be able to execute a tape, maintain the state,
and then execute a subsequent tape. This gives us an opportunity to add
runtime execution behaviour that is too complex to capture as a tape
operation (for example, because it relies heavily on runtime output).

We won't integrate this into the full E2E process here - this PR updates
the _backends_ to handle this type of execution. Changes to the pre- and
post-processing in the `FTQCQubit` device (and integration with
`convert_to_mbqc_formalism`) is the next step after this PR.

**Description of the Change:**

- We expand on the `execute` function of the `LightningQubitBackend` and
add a `simulate` function based on `lightning.qubit`'s `simulate`, but
limit it to shot-based executions only, and modify handling of state
reset and shot iteration to support multi-step execution.

- We add a `QuantumScriptSequence` object, to hope multiple
`QuantumScript`s that are part of a single execution, i.e. meant to be
executed in sequence before resetting the state (this differs from a
`QuantumScriptBatch`, which contains multiple quantum scripts each
representing a separate device execution).

- We add a function to split the tape at every non-Clifford gate (the
simplest implementation of this is fine for now)

**Benefits:**
We move one step closer to executing tapes in the MBQC formalism with
consolidated online corrections determined at runtime based on
measurement results.

**Possible Drawbacks:**
An alternative to this approach would be to investigate whether we can
insert a complex `qml.cond` (or a subclass of that that is its own
operation), and provide `lightning.qubit` with instructions for how to
handle it. This seemed simpler and faster with less mucking around in
updating existing architecture, so for a prototype, this strategy seemed
preferable.

[sc-93072][sc-93074]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>
Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>