[Unified Compiler] Integrate GraphStatePrepOp into the convert_to_mbqc_formalism pass (#8153)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
[sc-97745]

Recently, `mbqc.graph_state_prep` quantum ops and its decomposition pass
was added to the unified compiler module in this
[PR](https://github.com/PennyLaneAI/pennylane/pull/8090). This PR is a
following-up PR for the `convert_to_mbqc_formalism` pass as described
[here](https://github.com/PennyLaneAI/pennylane/blob/f85553c98878405d19f1e5460efc639e05093c5d/pennylane/compiler/python_compiler/transforms/convert_to_mbqc_formalism.py#L125).

Note that current implementation inserts only one type of adj matrix ops
in each region instead of one adj matrix for each gate operation. One
more optimal approach is to insert only one type of adj matrix ops as
needed for the whole module. As optimization is not the focus for now,
TODOs was added and we can come back to this in a following-up PR when
time permits.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>
Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>
Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>