`PrepSelPrep` dequeue its `lcu` input and has a single source of truth for its `lcu` (#8169)
**Context:**

`PrepSelPrep` has an input `lcu`, which is the linear combination of
unitaries that is being block-encoded.
Currently, `PrepSelPrep` stores the `coeffs` and `ops` of `lcu` but also
recreates `lcu` from `ops` and `coeffs` and stores this new linear
combination. This is not optimal, because
1. The reference to the input `lcu` is lost
2. The information about `lcu` is stored redundantly in `coeffs` + `ops`
and in `lcu`.

Additionally, `PrepSelPrep` currently does not de-queue its input `lcu`,
so that the following circuit queues two operators

```python
def circuit():
    lcu = qml.LinearCombination([0.25, 0.75], [qml.X(0), qml.Z(1)])
    qml.PrepSelPrep(lcu, control=["control"])
```

**Description of the Change:**
- Stop recreating `lcu` in `PrepSelPrep.__init__`.
- Make `PrepSelPrep.coeffs` and `PrepSelPrep.ops` inherit from
`PrepSelPrep.lcu` instead of storing them independently.
- De-queue the `lcu` input when queuing `PrepSelPrep` (this is made
possible by the first change)

**Benefits:**
Less redundancy/clutter in `PrepSelPrep`, and cleaner queuing behaviour.

**Possible Drawbacks:**
Tiny breaking change in the type of the stored `lcu` if it is not a
`LinearCombination`.

**Related GitHub Issues:**
This was split off of #8131 even though it is tightly related.~~,
because it is a breaking change to the UI.~~

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>