Remove usage of `pytest.mark.capture` from unified compiler tests (#8234)
The `pytest.mark.capture` marker is for tests that should be run by the
capture tests runner, but we want these tests to only run with the
external-tests runner. Just to be safe, this PR is removing the capture
marker and replacing it with explicit usage of the
`enable_disable_plxpr` fixture, which is added under the hood
automatically by the capture marker.