Determine mcm_method in setup_execution_config (#8184)
**Context:**

We are realizing there's issues with the current protocol of historical
pattern of:

1) handle mcms (defer measurements, dynamic one shot etc)
2) decompose

and the current pattern of

1) decompose
2) resolve dynamic wires
3) handle mcm's

Have issues.  

defer measurements may introduce new controlled objects that need to be
decomposed, but `decompose` may introduce new mcm's and wire
allocations.

**Description of the Change:**

Determine the mcm method in `setup_execution_config`. Organize the
preprocessing to more or less do:

1) defer measurements
2) decompose
3) resolve dynamic wires
4) dynamic one shot

We will later need to enforce that `decompose` does not introduce new
mcm's and conditionals if the mcm method is `deferred`.

Adds a `use_resets` kwarg to `resolve_dynamic_wires` and
`device_resolve_dynamic_wires`. This can be used to forbid the addition
of additional mcm's when `mcm_method == "deferred"`.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

**Related Shortcut Stories:**
[sc-98588]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>