`qml.Hermitian` renders incorrect results with `qml.probs` (#8235)
**Context:** `qml.Hermitian` and `qml.probs` are currently incompatible.
When combined, the output might be in incorrect order or totally
incorrect (see the examples in the linked issue). After a very long
discussion, it has been decided to just update the docs of both. We also
raise an explicit warning so that the user is certainly aware of this.

**Description of the Change:** As above.

**Benefits:** This behaviour is documented.

**Possible Drawbacks:** None that I can think of.

**Related GitHub Issues:** #6878

[sc-82976]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>