Implement ResourceConfig class (#8195)
**Context:**
The resource config currently exists as a global dictionary, which is
passed to plre.estimate
Issues:

- Not very user-friendly or sustainable, as the config will continue to
grow as more op-specific tolerances are added.
- From an SW standpoint, we also don’t want a global dict like this
- It also forces the user to make decisions about operators which they
do not necessarily care about, or to need to copy-paste all of the
defaults despite only being interested in changing one such parameter.
- The user needs to rewrite the config every time they want to change
something

**Description of the Change:**
Develops a config class, which has its own set of defaults, and makes it
easy for the user to change what they wish.

**Benefits:**
Resolves above issues.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-97227]

---------

Co-authored-by: Diksha Dhawan <40900030+ddhawan11@users.noreply.github.com>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Soran Jahangiri <40344468+soranjh@users.noreply.github.com>