Adding RC nightly builds (#8148)
**Context:**
The lates RC branch needs nightly builds.

**Description of the Change:**
Adding a workflow to build a python wheel from the latest RC branch and
upload it to testpypi. The workflow runs daily and checks if RC branch
exists then bumps version in the RC branch.

**Benefits:**
RC branch gets a nightly build.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>