Qnode primitive captures unprocessed `ExecutionConfig` (#8258)
**Context:**

Blocks https://github.com/PennyLaneAI/catalyst/pull/2041

**Description of the Change:**

The capture `Device.setup_execution_config` was about the capabilities
of the capture execution pipeline, which is no longer maintained or
actively developed. While I'm not sure where will be the best place for
this logic in catalyst, we can leave that up to catalyst and no
determine it in program capture.

**Benefits:**

Unblocks catalyst unification work.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-99193]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>