add support for measure and conditional to to_openqasm (#8210)
**Context:**

Feature request for the ability to use `to_openqasm` with `qml.measure`.

**Description of the Change:**

Adds conversion for `qml.measure` and `qml.cond`.

**Benefits:**

**Possible Drawbacks:**

I'm not entirely sure this is the right openqasm 2 syntax, as the
current documentation for version 2 is rather lacking.

But this paper seems to use the syntax I'm following:
https://arxiv.org/pdf/1707.03429

**Related GitHub Issues:**

Fixes #6900 [sc-83351]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>