Port non-parametric resource operators (#8240)
**Context:** Ports non-parametric resource operators to the `estimator`
module

**Description of the Change:** Following changes are done:
1. `qml_estimator.rst`: Updated to list all the key objects for the
module. Shifted stuff from `estimator/__init__.py` and consolidated them
here.
2. `conf.py`: Added two custom hooks to the Sphinx configuration. The
first one acts when stubs are read. It checks for `qml.estimator.ops` in
the path and attaches `:noindex:` directive to it. The second one acts
after the stubs have been written as HTML, and adds links to the
non-indexed stubs ONLY in the `qml_estimator` page.
3. `ops/identity.py`: contains `Identity` and `GlobalPhase` (yes, I know
this one is parametric, but it was easier to do).
4. `ops/qubit/non_parametric_ops.py`: contains all the non-parametric
resource operators.

**Benefits:**

**Possible Drawbacks:** only drawback for the new hooks for `sphinx`
configuration is that one won't be able to cross-reference operators
within the sub-module `qml.estimator.ops`. This side effect is limited
to that one sub-module and doesn't affect anything else; in fact, other
items in `qml.estimator` can still be cross-referenced.

**Related GitHub Issues:**

---------

Co-authored-by: Diksha Dhawan <40900030+ddhawan11@users.noreply.github.com>
Co-authored-by: Diksha Dhawan <ddhawan@umich.edu>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Austin Huang <austin.huang@xanadu.ai>
Co-authored-by: Soran Jahangiri <40344468+soranjh@users.noreply.github.com>
Co-authored-by: ANT0N <39093564+AntonNI8@users.noreply.github.com>