Fix `SpecialUnitary` Incompatibility at Torch (#8209)
**Context:**
The related issue #7583 reported a very suspicious bug that implies we
did not handle the interfaces of tensors properly for `SpecialUnitary`.
It did not involve any stress testing and the tested matrix is of quite
small size.

**Problem:**

The code path for more than 5 wires was hardcoded to use np.kron to
construct the Pauli basis matrices; and our pauli matrices there are
NumPy arrays (search `_pauli_matrices` in the src file). When these
NumPy arrays were multiplied by trainable Torch parameters, it triggered
an illegal implicit conversion of a `torch.Tensor` to a `numpy.ndarray`,
crashing the autograd engine.

**Description of the Change:**
Enforce the interface casting at the bugged place.
Add test for both torch and jax. autograd has been skipped since expm
not differentiable in vanilla autograd.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
https://github.com/PennyLaneAI/pennylane/issues/7583

[sc-92398]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>