Snapshots with tree-traversal and one-shot (#8140)
**Context:**

So far we've been ignoring snapshots with `one-shot` and
`tree-travsesal`, as each "snapshot" can correspond to many different
states.

**Description of the Change:**

The recorded measurement/ state can now correspond to a list of multiple
measurements, one for each time that snapshot is encountered.

**Benefits:**

**Possible Drawbacks:**

This could balloon in size for many shots or many mcm's. But the user
would just need to accept that.

**Related GitHub Issues:**

Fixes #8125  [sc-98188]

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>