`TransformContainer` stores original `TransformDispatcher` (#8306)
**Context:**

In order for capture to serve as only a frontend for catalyst and not
interfere with normal pennylane execution, we have to be able to turn on
program capture *after* the qnode definition. So the
`QNode.TransformProgram` needs to be able to hold on to the original
transform's primitive and plxpr transform.

While we could add those as additional arguments to
`TransformContainer`, the easier solution is just to make a breaking
change and have `TransformContainer` store the entire
`TransformDispatcher`. Honestly, this cleans up the whole transform
ecosystem and makes it much more maintainable as well.

**Description of the Change:**

Breaks `TransformContainer` to instead take `(transform :
TransformDispatcher, args: list|tuple = (), kwargs : dict | None=None)`.

**Benefits:**

Much simplified `TransformContainer`. It's just a transform with bound
args and kwargs.

This also continues setting us up to be able to implement
`TransformProgram.__call__` to be just:
```
def __call__(self, obj):
    for t in self:
       obj = t(obj, cotransform_cache=self.cotransform_cache)
   return obj
```

Though that will still take a couple more steps.

**Possible Drawbacks:**

Yes, this is a breaking change, but `TransformContainer` has been
properly kept internal so far. That makes it so that we can make these
bigger breaking changes without too many consequences.

**Related GitHub Issues:**

[sc-99741]

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>