[Draft] Add `heuristic_resources` attribute to `DecompositionRule` (#8293)
**Context:**
`DecompositionRule`s in the new decomposition system are assigned
resources. These resources are in principle allowed to be heuristically
estimated rather than exact, but we don't store whether this is the case
in the rule.

**Description of the Change:**
Store it in the rule as an attribute.

**Benefits:**
Can access information about the resources of the decomposition rule
from the decomposition rule, rather than quite indirectly from the test
suite...

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
N/A