Change ubuntu-latest to ubuntu-2404 (#8371)
**Context:**
Github will be updating their `ubuntu-latest` version from ubuntu-24 to
ubuntu-26 in the near future. This may cause issues later so it's best
to fix the ubuntu version and manually update it.

**Description of the Change:**
Changing `ubuntu-latest` to `ubuntu-24.04` on workflows that may be
version specific.

**Benefits:**
No unexpected issues when `ubuntu-latest` is updated in the future. 

**Possible Drawbacks:**
None. `ubuntu-latest` is currently ubuntu-24

**Related GitHub Issues:**