[v0.43.0 Release] Un-block CI by fixing `test_about.py` (#8416)
Our new nightly RC build workflow automatically updated v0.43.0's
version to 0.43.0-rc0 (see
[commit](https://github.com/PennyLaneAI/pennylane/commit/b4f1d80b40467756987993970ae57c80ba5537c5)).
This is technically valid but the test failed as it thinks this is a
`dev` prerelease and expects the wheels to be named `v0.43.0.rc0` and
not `v0.43.0rc0`.

Quick fix to the test to work with the current version name.

[sc-100943]