Fix diagonalize_measurements xdsl pass bug (#8383)
**Context:**

Trying to integrate the `diagonalize_measurements` pass with
`convert_to_mbqc_formalism` uncovered a bug in the diagonalization pass.
It was written assuming that the SSA value passed into the observable
would never be used again.

However, while it's true that additional _quantum_ operations won't be
applied, other operations regarding deallocation of qubits or
manipulation of registers can (and do) use the same SSA value. So when
we update the SSA value going into the observable, we can't update
_only_ the observable.

**Description of the Change:**
Update the qubit, everywhere where it's not being fed into a `CustomOp`
(since this is the place where it's used by the diagonalizing gate as
input), before changing the observable to the Z basis.

**Benefits:**
No more bug. We can use `diagonalize_measurements` and
`convert_to_mbqc_formalism` together in the capture pipeline.

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>