Be able to turn on capture after qnode definition (#8307)
**Context:**

In order for capture to serve as *only* a frontend for catalyst, we need
to be able to turn on capture *after* qnode definition.

This just slightly changes how qnode transformations get applied so they
are handled lazily.

Depends on #8306 

**Description of the Change:**

Qnode transformations are captured on call time from the transform
program, instead of transforming the qnode to be a function.

**Benefits:**

Can locally turn capture on and off inside of `qml.qjit`.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-99740]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>
Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>