Add jax-xdsl as a dependancy and updte stableHLO imports to use the new repo (#8465)
**Context:**
As xdsl migrated the upstream stableHLO dialect to a new
[repo](https://github.com/xdslproject/xdsl-jax), they removed it from
the main upstream xdsl repo. Therefore updating xdsl version to 0.54.0
or beyond would break trying to access the stableHLO dialect

**Description of the Change:**
Updated the imports to pull in the upstream dialect from the `xdsl-jax`
repo and hence adding it as a new dependency in PennyLane.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-101286]