[Deprecation] rename `argnum` to `argnums` in `qml.grad` and `qml.jacobian` (#8481)
**Context:**

We are moving our ecosystem closer and closer to the jax/ catalyst way
of doing things, instead of the autograd way of doing things. To make
`qml.grad` and `qml.jacobian` more compatible with the jax/ catalyst way
of doing things, we are renaming `argnum` to `argnums`.

**Description of the Change:**

Renames `argnum` to `argnums` and raises a deprecation warning if
`argnum` is provided instead.

**Benefits:**

Consistent naming through the ecosystem, leading to a more unified
pennylane and catalyst.

Also, turns out that lightning isn't using `argnum` in any of the tests,
so we don't have to worry about breaking it :)

**Possible Drawbacks:**

Is still a deprecation, and will probably break a bunch of ecosystem
tests.

**Related GitHub Issues:**

[sc-101569]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>