Rename `argnum` to `argnums` in vjp and jvp (#8496)
**Context:**

We are wanting to better unify Catalyst and PennyLane, and be better
able to use pennylane as the complete catlayst frontend.

While jax and Catalyst use `argnums` (plural) pennylane still uses
`argnum`. This makes it impossible to replace one with the other without
changes. Since we are wanting to be more jax-like than autograd-like
these days, we are proceeding with `argnums`.

**Description of the Change:**

Renames `argnum` to `argnums` and introduces a deprecation cycle for
`argnum`. Branches of a similar change for `grad` and `jacobian`.

**Benefits:**

Makes my life easier when updating catalyst tests to use capture as the
frontend.

**Possible Drawbacks:**

It is a deprecation cycle.

**Related GitHub Issues:**

[sc-101819]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>