Update documentation of GlobalPhase (#8521)
**Context:**
The existence of a controlled global phase is an inevitable intermediate
product of decompositions, but the concept is confusing, since the
control makes the "global phase" not "global" anymore. The user
shouldn't be explicitly creating a controlled global phase. We cannot
disallow it because that would break decompositions, but we should at
least stop advertising it.

**Description of the Change:**
Removes section in the `GlobalPhase` docstring about controlling a
`GloablPhase`

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-101986]