Fix bug in `QubitUnitaryOp` in xDSL quantum dialect. Add constructor tests (#8456)
**Context:**
The constructor method `__init__` of `QubitUnitaryOp` in the `Quantum`
dialect in the xDSL integration is using an undefined variable. This is
not caught because we nowhere test the constructor methods of operations
in that dialect.

**Description of the Change:**
Add unit tests for `__init__` of all operations in the quantum dialect.
Fix the undefined variable bug in `QubitUnitaryOp`.

**Benefits:**
Code/Test quality

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
[sc-101237]