Enable `cancel_inverses` to cancel nested pairs of mutually inverse gates (#8483)
**Context:**
`cancel_inverses` cancels one pair of gates at a time and then moves on
through the circuit. This means that nested pairs of mutually inverse
gates are not cancelled.

In addition, `cancel_inverses` as an almost-bug undesired behaviour
#6729.

**Description of the Change:**
Add keyword argument `recursive` that enables iterative cancellation of
such nested pairs. It is enabled by default and
can be disabled for backwards-compatible behaviour.

Note that repeatedly applying `cancel_inverses` can lead to the same
result, but it is not known a priori how many times it needs to be
called, and calling it repeatedly is less efficient.

In addition, this PR alters behaviour slightly to fix #6729.

to do:
- [x] discuss naming of kwarg.
- [x] ~~plxpr transform~~.

**Benefits:**
Better compilation pass quality.
"bug" fix

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
Fix #6729 
[sc-80821]