Capture PPR/PPM passes and register MLIR pass names (#8519)
**Context:**
Currently, we can run Catalyst/MLIR passes with the unified compiler
alongside xDSL passes, _if they are represented in the MLIR_. However,
for current MBQC workloads, we want to be able to use them the `capture`
enables, and currently they can't be captured (or converted to MLIR).



**Description of the Change:**
We add dummy versions of the transforms that just raise a
`NotImplemented` error if executed, but are `transform` objects and are
captured in `plxpr`. We register them with the relevant Catalyst pass
name so they can be converted from plxpr to MLIR.


**Benefits:**
We unblock current MBQC work with the PPR/PPM compilation pipeline.

**Possible Drawbacks:**
A more long-term solution to unifying the API for the various
transform/pass types is on the roadmap, and ideally we would use that
instead of manually add the transforms we are interested in to the
`ftqc` module. This is intended to be a temporary patch to unblock
current MBQC work.

---------

Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>