Gates in the gate set are not considered solved in the graph (#8526)
**Context:**
A `KeyError` is raised when trying to access the decomposition rule of
an operator that is in the gate set.

**Description of the Change:**
Now the graph correctly returns `False` when `is_solved_for` is called
on an op in the gate set, and the correct error is raised when trying to
access the decomposition rule of such an operator.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-102050]