Add `get_constant_from_ssa` function for getting the concrete value of an xDSL variable (#8514)
There are currently two ways of defining constant values -
`arith.constant`, and `stablehlo.constant` - and we need to deal with
them in different ways when trying to extract the concrete value. This
PR adds a function called `get_constant_from_ssa`, which takes an SSA
value as input, and returns its concrete value if it is a constant, else
`None`.

Notes:
* Added a `utils.py` file to the `python_compiler` submodule
* The function can currently only extract scalar values, and explicitly
returns `None` if the value is not scalar, even if it is a constant.

[sc-100957]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>