Stop using `import pennylane as qml` in devices module Pt 1 (#8459)
**Context:**

Now that we've changed how `TransformDispatcher` works, we can stop
using `import pennylane as qml` in the devices module. This sets us to
be be able to categorize it as a core module with tach.

**Description of the Change:**

Update all imports in the devices module.

**Benefits:**

Improved source code dependency handling.

**Possible Drawbacks:**

I think I'm just going to do some files in this PR instead of all of
them to avoid it getting too big.

**Related GitHub Issues:**

[sc-101268]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>