Add split-non-commuting pass (#8531)
**Context:**
We need a minimum working prototype of `split_non_commuting` to enable
EOY target workflows. We have competing priorities, so we have chosen to
limit the requirements here to just target the specific EOY goals; a
progressively more general version of the pass will have to be
implemented later.

Target workflows all return a Hadamard test, that is to say
`expval(X(0)), expval(Y(0))`.

This means this prototype is not required to add support for:
- `CompositeOp` observables
- `SymbolicOp` observables
- Additional `MeasurementProcesses` beyond `Expval`
- Efficient handling of duplicate observables
- Multiple options for grouping methods

Working with the following target circuit while preserving program
structure and returning the expected results is sufficient to meet the
requirements for this quarter:

```python
dev = qml.device("lightning.qubit", wires=3)

def qfunc(num_steps):
    qml.Hadamard(0)

    for i in range(num_steps):
        qml.Rot(0.12, 0.56, 0.78, 0)
        qml.Rot(2.3, 0.34, 0.67, 1)
        qml.Rot(1.2, 0.89, 1.9, 2)
        qml.ctrl(qml.MultiRZ(0.345, wires=[1, 2]), control=0)
        qml.PhaseShift(-0.567, wires=0)
        
@qml.qjit(pass_plugins=[getXDSLPluginAbsolutePath()], autograph=True)
@measurements_from_samples_pass
@diagonalize_final_measurements_pass
@split_non_commuting_pass
@qml.qnode(dev, shots=1000)
def circ(num_steps: int):
    qfunc(num_steps)
    return [(qml.expval(qml.PauliX(i)), qml.expval(qml.PauliY(i))) for i in range(3)]

circ(100)
```

[sc-93071]

---------

Co-authored-by: lillian542 <38584660+lillian542@users.noreply.github.com>
Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>