deprecation: `is_hermitian` -> `is_verified_hermitian` renaming of `Operator` property (#8494)
This property needs to be renamed to more accurate to its functionality.
We will now have the following behaviour,
```
>>> import pennylane as qml
>>> op = qml.X(0)
>>> op.is_hermitian
PennyLaneDeprecationWarning: The `is_hermitian` property is deprecated and has been renamed to `is_verified_hermitian` as it better reflects the functionality of this property. The deprecated access through `is_hermitian` will be removed in PennyLane v0.45. Alternatively, consider using the `pennylane.is_hermitian` function instead as it provides a more reliable check for hermiticity. Please be aware that it comes with a higher computational cost.

True
>>> op.is_verified_hermitian
True

```

Here's an example of the documentation rendered:
https://xanaduai-pennylane--8494.com.readthedocs.build/en/8494/code/api/pennylane.Hadamard.html#pennylane.Hadamard.is_hermitian